tests: disable test_255_bytes_null_end
Previously, that test wasn't run either.
I've fixed some buffer overflows and adjusted the test to do what I
think was the original intention.

The buffers do not match, yet though. The issue is that [256] is 0x01
instead of 0xAB. I assume that the intention here was to check if the
encoder writes anything unnecessary after the encoded data.
Unfortunately it seems to do just that though and I didn't yet spend
time into finding a fix for that.